migrar a BigDecimal y agregar Bean Validation para montos con max 2 decimales

+++ src/main/java/com/mendel/challenge/dto/SumResponse.java
@@ -3,9 +3,11 @@ package com.mendel.challenge.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class SumResponse {
 
-    private double sum;
+    private BigDecimal sum;
 }
